Pushing updated tests, see comments for set_point_device
Issues have been observed with set_point_device. The elec_coord_d
pointer seems to be correctly allocated from C then passed to the
Fortran driver. However, when switching back to C with the
set_point_device call, the coord pointer gets a different value
(resulting in a segfault on the GPU when trying to access it)